backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
pydantic[email]
python-dotenv
python-multipart
bcrypt
pyjwt
Pillow>=10.2.0
opencv-python-headless
numpy
requests
httpx
aiofiles
gunicorn
